requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
langgraph
pydantic
python-dotenv
streamlit
fastapi
uvicorn[standard]
google-genai
together
twilio
requests
duckduckgo-search
wikipedia
pillow
